improve group betweenness centrality (#4435)
* In this PR I improved the group betweenness centrality algorithm by using the algorithm of Puzis et al. "Fast algorithm for successive computation of group betweenness centrality".

In the previous implementation of the algorithm, for each pair of source - target non - group node permutation all the shortest paths were found and if one of them go through one of the nodes in the group then the group betweenness centrality increases. This was done for each group.

Here i used Pusiz et al. implementation, which finds the GBC with running time of $O(K^3)$ for each group (k is the number of nodes in the group) and with a preprocessing time of $O(mn)$ for all the groups.

This unable us to find the GBC of many groups in one preprocessing step.
 furthermore, the article shows how to find the prominent group of size k (the group with the highest GBC among those with size k), which i would add in future PR.

* In this PR I improved the group betweenness centrality algorithm by using the algorithm of Puzis et al. "Fast algorithm for successive computation of group betweenness centrality".

In the previous implementation of the algorithm, for each pair of source - target non - group node permutation all the shortest paths were found and if one of them go through one of the nodes in the group then the group betweenness centrality increases. This was done for each group.

Here i used Pusiz et al. implementation, which finds the GBC with running time of $O(K^3)$ for each group (k is the number of nodes in the group) and with a preprocessing time of $O(mn)$ for all the groups.

This unable us to find the GBC of many groups in one preprocessing step.
 furthermore, the article shows how to find the prominent group of size k (the group with the highest GBC among those with size k), which i would add in future PR.

* In this PR I improved the group betweenness centrality algorithm by using the algorithm of Puzis et al. "Fast algorithm for successive computation of group betweenness centrality".

In the previous implementation of the algorithm, for each pair of source - target non - group node permutation all the shortest paths were found and if one of them go through one of the nodes in the group then the group betweenness centrality increases. This was done for each group.

Here i used Pusiz et al. implementation, which finds the GBC with running time of $O(K^3)$ for each group (k is the number of nodes in the group) and with a preprocessing time of $O(mn)$ for all the groups.

This unable us to find the GBC of many groups in one preprocessing step.
 furthermore, the article shows how to find the prominent group of size k (the group with the highest GBC among those with size k), which i would add in future PR.

* In this PR I improved the group betweenness centrality algorithm by using the algorithm of Puzis et al. "Fast algorithm for successive computation of group betweenness centrality".

In the previous implementation of the algorithm, for each pair of source - target non - group node permutation all the shortest paths were found and if one of them go through one of the nodes in the group then the group betweenness centrality increases. This was done for each group.

Here i used Pusiz et al. implementation, which finds the GBC with running time of $O(K^3)$ for each group (k is the number of nodes in the group) and with a preprocessing time of $O(mn)$ for all the groups.

This unable us to find the GBC of many groups in one preprocessing step.
 furthermore, the article shows how to find the prominent group of size k (the group with the highest GBC among those with size k), which i would add in future PR.

* backwards-compatible with the previous implementation

* backwards-compatible with the previous implementation

* After review

* After review

* After review

* After review

* improving the runtime of the algorithm and fixing bug

* changing the help functions

* changing the help functions

* changing the help functions

* change the helper function and now the output value for a single group is an integer

* change the helper function and now the output value for a single group is an integer

* change the function description

* after review. Change normalize and docstring

* divide by 2 if G is undirected

* after changes