Only compute shortest path lengths when used (#5183)
In the case when alpha is 1, the expression making use of the path_len is
nullified by the preeceding factor evaluating to zero. Computing a dictionary of
all shortest path lengths is expensive. We should avoid computing this when the
result isn't used. 1 is a common enough value for alpha that the added
complexity of this optimization is sufficiently justified.