Update simple_paths.py: consistent behaviour for `is_simple_path` when path contains nodes not in the graph. (#6272)
* Update simple_paths.py

Current version of `is_simple_path` fails with `KeyError` if the first element of the node list is not in the graph.

* Update simple_paths.py

* Update simple_paths.py

Simplify the test condition. Checking for a single node in the list can be combined with checking for duplicates in the list without meaningful efficiency loss.

* Update test_simple_paths.py

Add the case when the start of the path is a node not in the graph.

* Update simple_paths.py

* Update simple_paths.py

* Update simple_paths.py

* Update simple_paths.py

* Update simple_paths.py

Still need to check the special case of a list with 1 item.